+++ NEWS.md
@@ -1,6 +1,13 @@
 
 # starschemar (development version)
 
+# starschemar 1.1.0
+* New dimension enrichment features: [enrich_dimension_export()] and [enrich_dimension_import()].
+* New functions to filter facts and purge dimensions (delete unnecessary data): [filter_fact_rows()] and [purge_dimensions()].
+* New query definition functions on the exported multistar format: [dimensional_query()], [select_fact()], [select_dimension()], [filter_dimension()] and [run_query()].
 
 # starschemar 1.0.0
-
+* Initial functionality for defining star schemas and constallations.
+* Data cleaning.
+* Incremental refresh.
+* Results export.